add angular cli frontend

+++ backend/src/main/java/scis/service/SCService.java
@@ -8,7 +8,7 @@ import java.util.Collection;
  * Created by wangyifei on 2017/6/11.
  */
 public interface SCService {
-    SC findById(Long id);
+    SC findById(int id);
     Collection<SC> findAllSC();
     void saveSC(SC sc);
     void deleteSC(SC sc);
